Fix editor import regression after Monaco migration
