+++ src/components/package-port/CodeEditorHeader.tsx
@@ -1,8 +1,8 @@
-import React, { useState, useCallback, useMemo } from "react"
+import React, { useState, useCallback } from "react"
 import { Button } from "@/components/ui/button"
 import { handleManualEditsImportWithSupportForMultipleFiles } from "@/lib/handleManualEditsImportWithSupportForMultipleFiles"
-import { useImportComponentDialog } from "@/components/dialogs/import-component-dialog"
 import { useToast } from "@/hooks/use-toast"
+import { useEditorComponentImport } from "@/hooks/use-editor-component-import"
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -32,14 +32,7 @@ import {
   TooltipProvider,
   TooltipTrigger,
 } from "@/components/ui/tooltip"
-import {
-  JlcpcbComponentTsxLoadedPayload,
-  KicadStringSelectedPayload,
-  TscircuitPackageSelectedPayload,
-} from "@tscircuit/runframe/runner"
 import { ICreateFileProps, ICreateFileResult } from "@/hooks/useFileManagement"
-import { useGlobalStore } from "@/hooks/use-global-store"
-import { openJlcpcbImportIssue } from "@/hooks/use-jlcpcb-component-import"
 export type FileName = string
 
 interface CodeEditorHeaderProps {
@@ -66,19 +59,15 @@ export const CodeEditorHeader: React.FC<CodeEditorHeaderProps> = ({
   createFile,
   aiAutocompleteState,
 }) => {
-  const { Dialog: ImportComponentDialog, openDialog: openImportDialog } =
-    useImportComponentDialog()
+  const { importComponentDialog, openImportDialog } = useEditorComponentImport({
+    currentFile,
+    files,
+    updateFileContent,
+    createFile,
+  })
   const { toast } = useToast()
   const [sidebarOpen, setSidebarOpen] = fileSidebarState
   const [aiAutocompleteEnabled, setAiAutocompleteEnabled] = aiAutocompleteState
-  const session = useGlobalStore((s) => s.session)
-
-  const jlcpcbProxyRequestHeaders = useMemo(() => {
-    if (!session?.token) return undefined
-    return {
-      Authorization: `Bearer ${session.token}`,
-    }
-  }, [session?.token])
 
   const handleFormatFile = useCallback(() => {
     if (!window.prettier || !window.prettierPlugins) return
@@ -168,116 +157,6 @@ export const CodeEditorHeader: React.FC<CodeEditorHeaderProps> = ({
       }
     }
   }, [currentFile, files, toast, updateFileContent])
-
-  const handleTscircuitPackageSelected = useCallback(
-    async ({ fullPackageName }: TscircuitPackageSelectedPayload) => {
-      if (!currentFile) {
-        const message = "Select a file before importing a component."
-        toast({
-          title: "No file selected",
-          description: message,
-          variant: "destructive",
-        })
-        throw new Error(message)
-      }
-
-      const existingContent = files[currentFile] ?? ""
-      const newContent = `import {} from "${fullPackageName}"\n${existingContent}`
-      updateFileContent(currentFile, newContent)
-      toast({
-        title: "Component imported",
-        description: `Added ${fullPackageName} to ${currentFile}.`,
-      })
-    },
-    [currentFile, files, toast, updateFileContent],
-  )
-
-  const handleJlcpcbComponentTsxLoaded = useCallback(
-    async ({ result, tsx }: JlcpcbComponentTsxLoadedPayload) => {
-      console.log(232, result, tsx)
-      const partNumber = result.component.partNumber || "component"
-
-      try {
-        const sanitizedBaseName = partNumber
-          .toLowerCase()
-          .replace(/[^a-z0-9_-]/gi, "-")
-        let componentPath = `imports/${sanitizedBaseName}.tsx`
-        let suffix = 1
-        while (files[componentPath] || files[`./${componentPath}`]) {
-          componentPath = `imports/${sanitizedBaseName}-${suffix}.tsx`
-          suffix += 1
-        }
-
-        const createFileResult = createFile({
-          newFileName: componentPath,
-          content: tsx,
-          onError: (error) => {
-            throw error
-          },
-        })
-
-        if (!createFileResult.newFileCreated) {
-          throw new Error("Failed to create component file")
-        }
-
-        toast({
-          title: "Component imported",
-          description: `${partNumber} saved to ${componentPath}.`,
-        })
-      } catch (error) {
-        const message =
-          error instanceof Error
-            ? error.message
-            : "Failed to import component from JLCPCB"
-
-        toast({
-          title: "JLCPCB import failed",
-          description: (
-            <div className="space-y-2">
-              <p>{message}</p>
-              <button
-                className="text-sm text-blue-500 hover:underline"
-                onClick={(event) => {
-                  event.preventDefault()
-                  openJlcpcbImportIssue(partNumber, message)
-                }}
-              >
-                File issue on GitHub
-              </button>
-            </div>
-          ),
-          variant: "destructive",
-        })
-
-        throw new Error(message)
-      }
-    },
-    [createFile, files, toast],
-  )
-
-  const handleKicadStringSelected = useCallback(
-    async ({ footprint, result }: KicadStringSelectedPayload) => {
-      try {
-        await navigator.clipboard.writeText(footprint)
-        toast({
-          title: "KiCad footprint copied",
-          description: `${result.footprint.qualifiedName} copied to clipboard.`,
-        })
-      } catch (error) {
-        const message =
-          error instanceof Error
-            ? error.message
-            : "Failed to copy KiCad footprint to clipboard"
-        toast({
-          title: "KiCad import failed",
-          description: message,
-          variant: "destructive",
-        })
-        throw new Error(message)
-      }
-    },
-    [toast],
-  )
 
   return (
     <>
@@ -481,12 +360,7 @@ export const CodeEditorHeader: React.FC<CodeEditorHeaderProps> = ({
             </DropdownMenu>
           </div>
         </div>
-        <ImportComponentDialog
-          onTscircuitPackageSelected={handleTscircuitPackageSelected}
-          onJlcpcbComponentTsxLoaded={handleJlcpcbComponentTsxLoaded}
-          onKicadStringSelected={handleKicadStringSelected}
-          jlcpcbProxyRequestHeaders={jlcpcbProxyRequestHeaders}
-        />
+        {importComponentDialog}
       </div>
     </>
   )
